useStudyDetailQuery 파일 변경
